python/step8-monitoring-3rd-party-libs/pyproject.toml
[project]
name = "my-container-infra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.172.0",
    "cdk-monitoring-constructs>=9.7.1",
    "constructs>=10.4.2",
    "pytest>=8.3.4",
]
